Reorganizing the code structure
